Feature/sscs 1862 send to number (#7)
In SendToNumber.js, add the field to the form function with validation and content. In the next function add branching functionality. If the user selects 'Yes` go to the sms-confirmation page. If 'No' then to the enter-mobile page.
Add the content for the field, question and error message in send-to-number content.json file.
Update the page template to render the field.

In TextReminders.js in template method, determine whether the number the user entered was a mobile number or not using regex. Then pass it to the locals. In the textreminders template do the logic for the sign up button href. If it is a mobile then the href = /send-to-number otherwise set the the href to /enter-mobile.

In regex.js adjust the phoneNumnber regex to allow numbers with + in then. e.g. +447127456987.

In steps.js Add two new steps, one for enter mobile and the other for sms confirmation.
For both enter-mobile and sms-confirmation, add directories in the sms-notify folder with placeholder content.json, template.html and javascript files.