SSCS-1292 Enter number (#10)
Add Enter number page to form.
In EnterMobile.js add the content, field, field validation and next step.
Add the page content to the content.json. Edit the template.html to add the field and content.
The next step is the sms-confirmation page.

In validation Utils add the ability to add multiple validation by using the multiRegex method.
Pass an array of objects with the regex and the error message to look for in the content.json.