SSCS-1863 SMS confirmation (#12)
Add SMS-confirmation page.
Add template with link to go back to the enter-mobile step and link to progress to the representative step. Display the number the user has entered. If mobile number has been entered in the /enter-mobile step then display that else display the appellant phone number entered in the appellant-details step.
Add content to the json file.