SSCS-2195 Postcode checker (#107)
Add bodyparser module to parse body from form to server.
Add valid-postcode directory.
In the routes.js for valid-postcode, add routes for /postcode-check (GET, POST) and /invalid-postcode.
For /postcode-check (POST) add logic to verify if postcode value is: A) valid and B) in the list. If valid and in the list then redirect to /benefit-type. If valid and NOT in the list redirect to /invalid-postcode. If invalid then add error handling logic and pass to the template.
Add E2E tests to test the routing logic
Improve postcode regex and add regex for getting the second part of the postcode.
Add content and template for postcodeCheck page
Add template for invalid postcode page