SSCS-2226: Revert extend from Page
extend from Question because currently it's not possible to retrieve an
answer from previous step from a Page type.

Add function benefitType to get the content for the wording in the page.

Add text in the content json file

call the benefitType function from the template to display the wording

Add unit test to cover production code