SSCS-2421: Add Enter your name input field
Add input field and wire it up in a way that it is posted together
with the appeal json message.

Add Values and Answers methods to the controller. Answers one is empty
since we don't need to display any information on the page.

Overwrite template block to customise it and display what I need