SSCS-2421: Debug and fix issue on the CheckYourAppeal page
The issue is that the CheckYourAppeal template was not displaying the
incomplete text accordingly. For example, when there were
incomplete steps the CheckYourAppeal page was displaying the complete
text to sign and submit the appeal.

It turns out the root cause was the empty Answers() methods in the
Questions classes. Since we were returning an empty array there was
no way to figure out whether the answer in the Step was completed and
by default it was returning it as incomplete.

The fix is to clean up classes of those empty Answers() and Values().