Feature/sscs 2421 sign and submit (#188)
* SSCS-2421 Add sign and submit text to template

Prettify the appeal json message to make easier to read
Overwrite the statement of truth block in the template to display
the text indicated in the ticket requirement

* SSCS-2421 Add terms and condition link to the section

Add e2e test for the section

* SSCS-2421: Add Enter your name input field

Add input field and wire it up in a way that it is posted together
with the appeal json message.

Add Values and Answers methods to the controller. Answers one is empty
since we don't need to display any information on the page.

Overwrite template block to customise it and display what I need

* SSCS-2421: Clean up unused code

Remove the answers method from class since it is not used

* SSCS-2421: Refactor template

customise only the template blocks involved in this ticket and remove
no customised blocks. In order for this change to work it was necessary
a look and feel project improvement.

More details here: https://github.com/hmcts/look-and-feel/pull/44/files

* SSCS-2421: Upgrade look-and-feel package

Update package to use latest improvements and to be able to
override the submit button in the check your answer template easily.

* SSCS-2421: Add unit test coverage to class

Add unit test coverage for the form and values methods
Minor refactor for the ReasonForAppealing testing code such as
extract variable to reuse it and correct typo in the filename

* SSCS-2421: Minor Refactor

Change temporary log to keep consistency with other similar logs.
Remove hiddenInput macro from template as we don't have to customise it.

* SSCS-2421: Update the test wording to capture better the scenario

* SSCS-2421: Move error object within field to the end to keep consistency

* SSCS-2421: Debug and fix issue on the CheckYourAppeal page

The issue is that the CheckYourAppeal template was not displaying the
incomplete text accordingly. For example, when there were
incomplete steps the CheckYourAppeal page was displaying the complete
text to sign and submit the appeal.

It turns out the root cause was the empty Answers() methods in the
Questions classes. Since we were returning an empty array there was
no way to figure out whether the answer in the Step was completed and
by default it was returning it as incomplete.

The fix is to clean up classes of those empty Answers() and Values().

* SSCS-2421: Clean up CheckYourAppeal class

Remove method used to debug the issue
Remove commented out method

* SSCS-2421: Rewrite e2e scenario for the CheckYourAppeal page

I rewrite the scenario to test that the issue is fixed

* SSCS-2421: Fix e2e test

Update test to see the Sign and submit section

* SSCS-2421: Hide Other answers section on the CYA page

Return an Answer with the hide flag to true so that we hide those
questions on the Check Your Appeal page.

Return the values() as an empty array to clean up the json file from
values that we are not interested in.

* SSCS-2421: Add unit test for the answers method

Add unit test coverage for the answer.hide method
Clean up debug method

* SSCS-2421: Change regex to allow to sing with x

* SSCS-2421: Move content into the content file for i18n purpose

* Added spaces for consistency with the rest of the application

* SSCS-2421: Fix issue about displaying Submit and Sign section incorrectly

Issue: when reached certain steps in the journey and then going to the
/check-your-appeal the Sign and Submit your appeal is displayed
incorrectly as there are still required steps to complete

Root causes:
1. Incorrect section definition in some Step's answers.
2. No textfield required present in the form which is necessary for the
framework to figure out whether the step is complete or not.

* SSCS-2421: Changes from code review

Make OtherReasonForAppealing optional
Make ReasonForAppealing required and make sure that carriage returns
cannot validate this step.