Reverting Codecept.js config settings to their defaults
Due to a previous commit: 547d1e793a7f2636ba6209c49889adf451a10b7c the
time taken to run our functional tests has increased from 13 minutes to
40 minutes which is an increase of 207% - Great!

The current Nightmare.js default for waitForTimeout is 1000 this was
modified in the above commit to 10,000.

The current Nightmare.js default for waitForAction is 500 this was
modified in the above commit to 2000.

This commit reverts these values back to their defaults ensuring our
tests run in a timely manner.