Public key pinning
pin-sha256 - The quoted string is the Base64 encoded Subject Public
Key Information (SPKI) fingerprint. It is possible to specify multiple
pins for different public keys. Some browsers might allow other hashing
algorithms than SHA-256 in the future. See below on how to extract this
information out of a certificate or key file.

max-age - The time, in seconds, that the browser should remember that
this site is only to be accessed using one of the defined keys.