SSCS-3266 Evidence reminder (#446)
Fixed a bug where one-per-page would skip the Evidence Reminder page. This is because previously it was using the form method extended from the Question class. This was so we could get a value from another step. However, one-per-page has some logic where it says: if the page has been validated and is valid then move on the next step (I believe this is to do with changing your answer in the check your answers page). As there is no actual field on this page and is just content, it was always valid and so was skipped.
In SendingEvidence.js

Changed the class to extend from Interstitial class not Question.
Removed form and answer method from class as it is not a part of Interstitial class.
Modified hasSignedUpForEmail getter to get the value of the field from the other step.
Also:

Modified the content and html of the page.
Modified the unit tests to reflect the class change using proxyquire to mock getting the value of the field from another step.